Move Nido Flask server helper functions to separate file
